+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- *Breaking* Changed how backup details are stored in the repository to 
+- *Breaking Change*:
+  Changed how backup details are stored in the repository to
   improve memory usage (#1735) from [vkamra](https://github.com/vkamra)
 
 ## [v0.0.3] (alpha) - 2022-12-05